--- iop/system/udnl-t300/src/IOPRP_img.c
@@ -1,8 +1,0 @@
-#ifndef __IOPRP_img__
-#define __IOPRP_img__
-
-unsigned int size_IOPRP_img = 0;
-unsigned char IOPRP_img[] __attribute__((aligned(16))) = {
-};
-
-#endif
