--- iop/system/udnl-t300/src/imports.lst
@@ -1,44 +1,0 @@
-dmacman_IMPORTS_start
-I_dmac_set_dpcr
-I_dmac_set_dpcr2
-I_dmac_set_dpcr3
-dmacman_IMPORTS_end
-
-intrman_IMPORTS_start
-I_CpuDisableIntr
-intrman_IMPORTS_end
-
-ioman_IMPORTS_start
-I_open
-I_read
-I_close
-I_lseek
-ioman_IMPORTS_end
-
-loadcore_IMPORTS_start
-I_GetLoadcoreInternalData
-I_QueryBootMode
-loadcore_IMPORTS_end
-
-stdio_IMPORTS_start
-I_printf
-stdio_IMPORTS_end
-
-sysclib_IMPORTS_start
-I_memset
-I_strcmp
-I_strncmp
-I_memcpy
-sysclib_IMPORTS_end
-
-sysmem_IMPORTS_start
-I_AllocSysMemory
-I_QueryBlockSize
-I_QueryBlockTopAddress
-I_QueryMemSize
-sysmem_IMPORTS_end
-
-thbase_IMPORTS_start
-I_SleepThread
-thbase_IMPORTS_end
-
