--- iop/system/udnl-t300/src/irx_imports.h
@@ -1,16 +1,0 @@
-#ifndef IOP_IRX_IMPORTS_H
-#define IOP_IRX_IMPORTS_H
-
-#include <irx.h>
-
-/* Please keep these in an alphabetical order!  */
-#include <dmacman.h>
-#include <intrman.h>
-#include <ioman.h>
-#include <loadcore.h>
-#include <stdio.h>
-#include <sysclib.h>
-#include <sysmem.h>
-#include <thbase.h>
-
-#endif /* IOP_IRX_IMPORTS_H */
